+++ Chirp.Core/CheepDto.cs
@@ -1,0 +1,2 @@
+namespace Chirp.Core;
+public record CheepDto(string Text, string Author, DateTime TimeStamp);
