+++ Chirp.Infrastructure/Author.cs
@@ -1,0 +1,7 @@
+public class Author
+{
+    public int AuthorId { get; set; }
+    public required string Name { get; set; }
+    public required string Email { get; set; }
+    public List<Cheep>? Cheeps { get; set; }
+}
