+++ Chirp.Infrastructure/Cheep.cs
@@ -1,0 +1,8 @@
+public class Cheep
+{
+    public int CheepId { get; set; }
+    public int AuthorId { get; set; }
+    public required Author Author { get; set; }
+    public required string Text { get; set; }
+    public DateTime TimeStamp { get; set; }
+}
