+++ Chirp.Infrastructure/CheepRepository.cs
@@ -1,5 +1,8 @@
+namespace Chirp.Infrastructure;
+
 using System.Linq.Expressions;
 using Microsoft.EntityFrameworkCore;
+using Chirp.Core;
 
 public class CheepRepository : ICheepRepository
 {
@@ -8,43 +11,26 @@ public class CheepRepository : ICheepRepository
     public CheepRepository(ChirpContext context)
     {
         _context = context;
-        DbInitializer.SeedDatabase(_context);
     }
 
-    public async Task<IEnumerable<Cheep>> GetCheeps(int pageIndex, int pageRange)
+    public async Task<IEnumerable<CheepDto>> GetCheeps(int pageIndex, int pageRange)
     {
         return await _context.Cheeps
             .Include(c => c.Author)
             .Skip((pageIndex - 1) * pageRange)
             .Take(pageRange)
+            .Select(c => new CheepDto(c.Text, c.Author.Name, c.TimeStamp))
             .ToListAsync();
     }
 
-    public async Task<IEnumerable<Cheep>> GetCheepsFromAuthor(string author, int pageIndex, int pageRange)
+    public async Task<IEnumerable<CheepDto>> GetCheepsFromAuthor(string author, int pageIndex, int pageRange)
     {
         return await _context.Cheeps
             .Include(c => c.Author)
             .Where(c => c.Author.Name == author)
             .Skip((pageIndex - 1) * pageRange)
             .Take(pageRange)
+            .Select(c => new CheepDto(c.Text, c.Author.Name, c.TimeStamp))
             .ToListAsync();
-    }
-
-    public async Task<Cheep?> Get(Expression<Func<Cheep, bool>> predicate)
-    {
-        return await _context.Cheeps
-            .Where(predicate)
-            .FirstOrDefaultAsync();
-    }
-
-    public async Task<IEnumerable<Cheep>> GetAll()
-    {
-        return await _context.Cheeps.ToListAsync();
-    }
-
-    public void Add(Cheep cheep)
-    {
-        _context.Cheeps.Add(cheep);
-        _context.SaveChanges();
     }
 }
