+++ Chirp.Infrastructure/Model.cs
@@ -1,3 +1,5 @@
+namespace Chirp.Infrastructure;
+
 using Microsoft.EntityFrameworkCore;
 
 public class ChirpContext : DbContext
@@ -16,21 +18,4 @@ public class ChirpContext : DbContext
 
     protected override void OnConfiguring(DbContextOptionsBuilder options)
         => options.UseSqlite($"Data source={DbPath}");
-}
-
-public class Author
-{
-    public int AuthorId { get; set; }
-    public required string Name { get; set; }
-    public required string Email { get; set; }
-    public List<Cheep>? Cheeps { get; set; }
-}
-
-public class Cheep
-{
-    public int CheepId { get; set; }
-    public int AuthorId { get; set; }
-    public required Author Author { get; set; }
-    public required string Text { get; set; }
-    public DateTime TimeStamp { get; set; }
 }
