+++ Chirp.Razor/Pages/UserTimeline.cshtml.cs
@@ -5,17 +5,17 @@ namespace Chirp.Razor.Pages;
 
 public class UserTimelineModel : PageModel
 {
-    private readonly ICheepService _service;
-    public List<CheepViewModel> Cheeps { get; set; }
+    private readonly ICheepRepository _cheepRepository;
+    public List<CheepDto> Cheeps { get; set; }
 
-    public UserTimelineModel(ICheepService service)
+    public UserTimelineModel(ICheepRepository cheepRepository)
     {
-        _service = service;
-        Cheeps = new List<CheepViewModel>();
+        _cheepRepository = cheepRepository;
+        Cheeps = new List<CheepDto>();
     }
     public async Task<IActionResult> OnGet(string author, [FromQuery(Name = "page")] int pageIndex = 1)
     {
-        var cheeps = await _service.GetCheepsFromAuthor(author, pageIndex);
+        var cheeps = await _cheepRepository.GetCheepsFromAuthor(author, pageIndex, 32);
         Cheeps = cheeps.ToList();
         return Page();
     }
