--- Chirp.Razor/Repository/IAuthorRepository.cs
@@ -1,5 +1,0 @@
-using Chirp.Razor.Repository;
-
-public interface IAuthorRepository : IRepository<Author> {
-
-}
