+++ Chirp.Razor/usings.cs
@@ -1,0 +1,2 @@
+global using Chirp.Core;
+global using Chirp.Infrastructure;
